Clean up diagnostic and landing page
