+++ pages/api/email.ts
@@ -9,14 +9,14 @@ export default async function handler(
   res: NextApiResponse
 ) {
   sgMail.setApiKey(process.env.SENDGRID_API_KEY);
-  const skills = req.body.skills;
+  const worksheets = req.body.worksheets;
   const msg = {
     to: req.body.email, // Change to your recipient
     from: process.env.SENDGRID_SENDER, // Change to your verified sender
     subject: "Your Math Champ Diagnostic Results",
     text: "and easy to do anywhere, even with Node.js", // for restrictive email clients
     html: "<strong>and easy to do anywhere, even with Node.js</strong>", // for email clients that can render CSS and HTML
-    attachments: skills.map((it) => {
+    attachments: worksheets.map((it) => {
       const pathToAttachment = `public${it.pdf}`;
       const data_base64 = base64_encode(pathToAttachment);
       return {
